Added code for "catch up" data delivery to processors.
Handles the case where a reference is added to shared data loader after
data has already been loaded. The loader sharing logic now saves its own
copy of loaded data. If data is already present when a reference is
added, it does a "catch up" delivery of the saved data to the added
processor.